2021 day 8: Extension
Prior to this change, Ben Lomax (@LomaxOnTheRun) wrote an amazing 13
line solution to today in Python.  The set work involved inspired my to
replicate it in Go.

This change adds set intersection and union functionality to the
RuneSets implemented for today, then applies Ben's set logic (slighlty
less concisely than his Python) instead of my original logic.

It should be noted that I didn't go to the effort of working out how to
make the RuneSets able to be frozen (and so be used as a map key), so
the matching of output digits is still done by scan :-( but that is
another level!